Refactor resultbackend and related models to separate packages.
